refactor: move table cell display function to helpers
