+++ src/helpers/getDisplayedValue/definedHeaders/customerHeaders.ts
@@ -1,0 +1,8 @@
+export default [
+  { text: "Name", prop: "name", type: "string" },
+  { text: "created At", prop: "createdAt", type: "dateTime" },
+  { text: "created By", prop: "createdByName", type: "string" },
+  { text: "updated At", prop: "lastModifiedAt", type: "dateTime" },
+  { text: "updated By", prop: "lastModifiedByName", type: "string" },
+  { text: "sap code", prop: "sapCode", type: "number" },
+];
