+++ src/helpers/getDisplayedValue/definedHeaders/index.ts
@@ -1,0 +1,3 @@
+import customerHeaders from "./customerHeaders";
+
+export default { customerHeaders };
