+++ src/interfaces/anyDataObject.ts
@@ -1,0 +1,4 @@
+export default interface IAnyDataObject {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  [key: string]: any;
+}
