+++ src/interfaces/tableHeader.ts
@@ -1,0 +1,5 @@
+export default interface ITableHeader {
+  text: string;
+  prop: string;
+  type: string;
+}
